feat(front-kit-3): extracted more order components
