+++ packages/falcon-ui-kit/src/Order/index.ts
@@ -3,3 +3,6 @@ export * from './OrderItemSummary';
 export * from './OrderItemSummaryLayout';
 export * from './EmptyOrderList';
 export * from './OrderListLayout';
+export * from './OrderListItemLayout';
+export * from './OrderListHeaderLayout';
+export * from './OrderListHeader';
